gpsd: disable C++ binding
Add flag to disable the C++ binding via EXTRA_OESCONS in the gpsd
bbappend.  The C++ binding is problematic as it seems to currently
be resulting in unresolved symbols at runtime unless libstdc++ is
linked in by the user.  Disable it for now since there are no known
users.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Id09c5cd434e1c6353ee75c7af2244b021f8b999f