Prepare master for new framework integration
During the last workshop the transition to the new framework was presented.
This change essentially deprecates the SMACK-based application framework.

To prepare the integration of it, we remove the deprecated components:
- recipes-apis : all agl-service-*
- recipes-demos: all demo apps but homescreen
- non-essential configurations

Todo:
We need to bring-back WAM/Chromium in a follow-up w/o ties into the old framework.

Bug-AGL: SPEC-4121

Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I58075f5f1e477106a1aad3d2171a0904bf0cb16e