tbtnavi: Bump SRCREV
Includes the following change:

- main: Wait until the channel is ready

This change is needed to make sure we are actually placed on the proper
output. This change is similar to the receiver side where are doing
something alike to make sure the gRPC change is in ready state.

Change-Id: I242c016a91468853c581c009e23946c0dce366ed
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/30739
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Scott Murray <scott.murray@konsulko.com>