rtl-sdr: update to latest release
Update rtl-sdr recipe to tagged 2.0.2 release, and switch to the
new CMake support, as it seems to be the focus of upstream maintenance
now.  A small patch is needed to tweak the installed library version
to match the tagged release version.

Note that the library now has usbfs zero-copy support, and that has
been enabled by default in the new PACKAGECONFIG support in the recipe
for now.  If any issues are reported that may be changed.

Bug-AGL: SPEC-5427

Change-Id: I0611bd1bc4e822c2966a41214e618a0c30f199fb
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/30944
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account