agl-service-*: Fix gateway demo configuration
The configuration parsing code for agl-service-hvac and
agl-service-audiomixer assumed that the Boost property_tree
parser would unquote strings, which is not the case.  This
broke overriding the TLS server name in the gateway demo
configuration files.  We can just not quote the strings for
now to avoid the problem.

This is effectively a Trout-specific workaround, as the
configuration parsing was replaced on master...

Bug-AGL: DEMO-ONLY!

Change-Id: Id94793932208f78186354b8d0891c5f94f649168
Signed-off-by: Scott Murray <scott.murray@konsulko.com>