Change the map from the Qt6 OSM plugin to MapLibre with OSM.
When Qt upgrades from 5 to 6, it droped the mapbox-gl plugin
from QtLocations.   After that, AGL uses the osm plugin.
It shows the "API Key Required" message on each map tile.
It could not erase.
Maplibre is an alternative solution for the mapbox-gl plugin
and the osm plugin for Qt6.

This patch enables the maplibre-native-qt plugin for QtLocations.

Bug-AGL: SPEC-5540

Change-Id: I28cd30344b2a846b185117afba20a881773d458c
Signed-off-by: Naoto Yamaguchi <naoto.yamaguchi@aisin.co.jp>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl-demo/+/31358
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
ci-image-build: Jenkins Job builder account
Tested-by: Jenkins Job builder account